task_manager/requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.6.0
apturl==0.5.2
asgiref==3.8.1
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
bson==0.5.10
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info===1.1build1
Django==5.1.1
django-admin==2.0.2
django-excel-response2==3.0.6
django-filter==24.3
django-rest==0.8.7
django-six==1.0.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
duplicity==0.8.21
email_validator==2.2.0
excel-base==1.0.4
exceptiongroup==1.2.2
fastapi==0.115.0
fastapi-cli==0.0.5
fasteners==0.14.1
future==0.18.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.20.2
httptools==0.6.1
httpx==0.27.2
idna==3.3
importlib-metadata==4.6.4
isoweek==1.3.3
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
markdown-it-py==3.0.0
MarkupSafe==2.0.1
mdurl==0.1.2
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
orjson==3.10.7
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
pydantic==2.9.2
pydantic-extra-types==2.9.0
pydantic-settings==2.5.2
pydantic_core==2.23.4
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
pymongo==4.9.1
PyMySQL==1.1.1
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-dotenv==1.0.1
python-multipart==0.0.10
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
rich==13.8.1
screen==1.0.1
SecretStorage==3.3.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
ssh-import-id==5.11
starlette==0.38.6
systemd-python==234
TimeConvert==3.0.13
typer==0.12.5
typing_extensions==4.12.2
tzlocal==5.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
ujson==5.10.0
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.30.6
uvloop==0.20.0
wadllib==1.3.6
watchfiles==0.24.0
websockets==13.1
xdg==5
xkit==0.0.0
xlwt==1.3.0
zipp==1.0.0
